Move permission tests to existing location
